frontend/core/blocks/crop-image-region-20241223-163051-184/computations.py
def compute(image_path, coordinates):
    '''
    Crops a specific region from an image based on the provided coordinates.
    
    Parameters:
    image_path (str): The file path to the image from which a region will be cropped.
    coordinates (dict): A dictionary containing the coordinates for cropping. It includes 'x1', 'y1' for the top-left corner and 'x2', 'y2' for the bottom-right corner of the rectangle.

    Returns:
    dict: A dictionary with the key 'cropped_image_path' and the value being the file path to the cropped image.
    '''
    from PIL import Image
    
    # Open the image file
    with Image.open(image_path) as img:
        # Convert coordinate values to integers
        x1 = int(coordinates['x1'])
        y1 = int(coordinates['y1'])
        x2 = int(coordinates['x2'])
        y2 = int(coordinates['y2'])
        
        # Crop the image using the provided coordinates
        cropped_img = img.crop((x1, y1, x2, y2))
        
        # Define the output path for the cropped image
        output_image_path = 'cropped_image.png'
        
        # Save the cropped image
        cropped_img.save(output_image_path)
    
    return {'cropped_image_path': output_image_path}

def generate_fake_image():
    from PIL import Image
    import numpy as np
    
    # Create a fake image (100x100 with 3 color channels)
    fake_image_array = np.random.randint(0, 255, (100, 100, 3), dtype=np.uint8)
    fake_image = Image.fromarray(fake_image_array)
    
    # Save this fake image as 'fake_image.png'
    fake_image_path = 'fake_image.png'
    fake_image.save(fake_image_path)
    
    return fake_image_path

def generate_coordinates():
    # Generate fake coordinates to test cropping
    return {'x1': 10, 'y1': 10, 'x2': 50, 'y2': 50}

def test_compute():
    fake_image_path = generate_fake_image()
    coordinates = generate_coordinates()
    
    result = compute(fake_image_path, coordinates)
    
    # Print result to verify
    print(result)
