frontend/core/blocks/crop-image-region-20241223-163051-184/requirements.txt
numpy==1.26.4
Pillow==11.0.0